Clean up delete app code
