+++ package.json
@@ -31,6 +31,7 @@
         "onCommand:appService.Start",
         "onCommand:appService.Stop",
         "onCommand:appService.Restart",
+        "onCommand:appService.Delete",
         "onCommand:appService.DeployZipPackage",
         "onCommand:appService.ZipAndDeploy",
         "onCommand:deploymentSlot.SwapSlots",
