--- src/explorer/appServiceNodes.ts
@@ -1,206 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { TreeDataProvider, TreeItem, TreeItemCollapsibleState, EventEmitter, Event, OutputChannel, window, MessageItem, MessageOptions } from 'vscode';
-import { AzureAccountWrapper } from '../azureAccountWrapper';
-import { SubscriptionClient, SubscriptionModels } from 'azure-arm-resource';
-import WebSiteManagementClient = require('azure-arm-website');
-import * as WebSiteModels from '../../node_modules/azure-arm-website/lib/models';
-import { NodeBase } from './nodeBase';
-import { DeploymentSlotNode } from './deploymentSlotNodes';
-import { DeploymentSlotsNode } from './deploymentSlotsNodes';
-import { FilesNode } from './filesNodes';
-import { WebJobsNode } from './webJobsNodes';
-import * as path from 'path';
-import * as opn from 'opn';
-import * as util from '../util';
-import * as kuduApi from 'kudu-api';
-
-export class SubscriptionNode extends NodeBase {
-    constructor(readonly subscription: SubscriptionModels.Subscription) {
-        super(subscription.displayName);
-    }
-
-    getTreeItem(): TreeItem {
-        return {
-            label: this.label,
-            collapsibleState: TreeItemCollapsibleState.Collapsed,
-            contextValue: 'subscription',
-            iconPath: { 
-                light: path.join(__filename, '..', '..', '..', '..', 'resources', 'light', 'AzureSubscription_16x.svg'),
-                dark: path.join(__filename, '..', '..', '..', '..', 'resources', 'dark', 'AzureSubscription_16x.svg')
-            }
-        }
-    }
-
-    async getChildren(azureAccount: AzureAccountWrapper): Promise<NodeBase[]> {
-        if (azureAccount.signInStatus !== 'LoggedIn') {
-            return [];
-        }
-
-        const credential = azureAccount.getCredentialByTenantId(this.subscription.tenantId);
-        const client = new WebSiteManagementClient(credential, this.subscription.subscriptionId);
-        const webApps = await client.webApps.list();
-        const nodes = webApps.sort((a, b) => {
-            let n = a.resourceGroup.localeCompare(b.resourceGroup);
-
-            if (n !== 0) {
-                return n;
-            }
-
-            return a.name.localeCompare(b.name);
-        }).map<AppServiceNode>((site, index, array) => {
-            return new AppServiceNode(site, this.subscription);
-        });
-
-        return nodes;
-    }
-}
-
-export class AppServiceNode extends NodeBase {
-    constructor(readonly site: WebSiteModels.Site, readonly subscription: SubscriptionModels.Subscription) {
-        super(site.name);
-    }
-
-    getTreeItem(): TreeItem {
-        if (!this.site.kind.startsWith('functionapp')) {
-            let iconName =  'AzureWebsite_16x_vscode.svg';
-            return {
-                label: `${this.label} (${this.site.resourceGroup})`,
-                collapsibleState: TreeItemCollapsibleState.Collapsed,
-                contextValue: 'appService',
-                iconPath: { 
-                    light: path.join(__filename, '..', '..', '..', '..', 'resources', 'light', iconName),
-                    dark: path.join(__filename, '..', '..', '..', '..', 'resources', 'dark', iconName)
-                }
-            }
-        }
-    }
-
-    async getChildren(azureAccount: AzureAccountWrapper): Promise<NodeBase[]> {
-        if (azureAccount.signInStatus !== 'LoggedIn') {
-            return [];
-        }
-        
-        // https://github.com/Microsoft/vscode-azureappservice/issues/45
-        return [
-            new DeploymentSlotsNode(this.site, this.subscription),
-            // new FilesNode('Files', '/site/wwwroot', this.site, this.subscription),
-            // new FilesNode('Log Files', '/LogFiles', this.site, this.subscription),
-            new WebJobsNode(this.site, this.subscription)
-        ];
-    }
-
-    browse(): void {
-        const defaultHostName = this.site.defaultHostName;
-        const isSsl = this.site.hostNameSslStates.findIndex((value, index, arr) => 
-            value.name === defaultHostName && value.sslState === "Enabled");
-        const uri = `${isSsl ? 'https://' : 'http://'}${defaultHostName}`;
-        opn(uri);
-    }
-
-    openInPortal(): void {
-        const portalEndpoint = 'https://portal.azure.com';
-        const deepLink = `${portalEndpoint}/${this.subscription.tenantId}/#resource${this.site.id}`;
-        opn(deepLink);
-    }
-
-    async start(azureAccount: AzureAccountWrapper): Promise<void> {
-        await this.getWebSiteManagementClient(azureAccount).webApps.start(this.site.resourceGroup, this.site.name);
-        return util.waitForWebSiteState(this.getWebSiteManagementClient(azureAccount), this.site, 'running');
-    }
-
-    async stop(azureAccount: AzureAccountWrapper): Promise<void> {
-        await this.getWebSiteManagementClient(azureAccount).webApps.stop(this.site.resourceGroup, this.site.name);
-        return util.waitForWebSiteState(this.getWebSiteManagementClient(azureAccount), this.site, 'stopped');
-    }
-
-    async restart(azureAccount: AzureAccountWrapper): Promise<void> {
-        await this.stop(azureAccount);
-        return this.start(azureAccount);
-    }
-
-    async delete(azureAccount: AzureAccountWrapper): Promise<boolean> {
-        return new Promise<boolean>(async (resolve, reject) => {
-            let servicePlanName = this.site.serverFarmId.substring(this.site.serverFarmId.lastIndexOf('/serverfarms/') + '/serverfarms/'.length);
-            let servicePlanRG = this.site.serverFarmId.substring(this.site.serverFarmId.indexOf('resourceGroups/') + 'resourceGroups/'.length, this.site.serverFarmId.indexOf('/providers/'));
-            let servicePlan = await this.getWebSiteManagementClient(azureAccount).appServicePlans.get(servicePlanRG, servicePlanName);
-
-            //for demoing purposes
-            console.log(await this.getWebSiteManagementClient(azureAccount).appServicePlans.list());
-    
-            let options = {};
-            options['deleteEmptyServerFarms'] = false;
-            let mOptions: MessageOptions = {modal: true};
-            console.log(servicePlan);
-            if (servicePlan.numberOfSites < 2) {
-                window.showInformationMessage(`This is the last web app on the plan, "${servicePlanName}".  Delete the Web App and App Service Plan?`, mOptions, ...['Both', 'Web App ONLY'])
-                .then(async (input) => {
-                    let deleteServicePlan = false;
-                    if (!input) {
-                        //if undefined, the user canceled the delete
-                        reject(false);
-                    } else  {
-                        if (input === 'Both') {
-                            deleteServicePlan  = true;
-                        }
-                        await this.getWebSiteManagementClient(azureAccount).webApps.deleteMethod(this.site.resourceGroup, this.site.name, {deleteEmptyServerFarm: deleteServicePlan});  
-                        resolve(true);
-                    }
-                });
-            } else {
-                window.showInformationMessage(`Delete the Web App "${this.site.name}"?`, mOptions, ...['Confirm'])
-                .then(async (input) => {
-                    if (!input) {
-                        resolve(false);
-                    } else {
-                        await this.getWebSiteManagementClient(azureAccount).webApps.deleteMethod(this.site.resourceGroup, this.site.name, ((error, response) => {
-                            console.log(response);
-                            resolve(true);
-                        }));
-                        
-                    }
-                    
-                });
-            }
-
-        });
-        
-    }
-
-    private getWebSiteManagementClient(azureAccount: AzureAccountWrapper) {
-        return new WebSiteManagementClient(azureAccount.getCredentialByTenantId(this.subscription.tenantId), this.subscription.subscriptionId);
-    }
-}
-
-export class NotSignedInNode extends NodeBase {
-    constructor() {
-        super('Sign in to Azure...');
-    }
-
-    getTreeItem(): TreeItem {
-        return {
-            label: this.label,
-            command: {
-                title: this.label,
-                command: util.getSignInCommandString()
-            },
-            collapsibleState: TreeItemCollapsibleState.None
-        }
-    }
-}
-
-export class LoadingNode extends NodeBase {
-    constructor() {
-        super('Loading...');
-    }
-
-    getTreeItem(): TreeItem {
-        return {
-            label: this.label,
-            collapsibleState: TreeItemCollapsibleState.None
-        }
-    }
-}
